chore: Rename internal module
