+++ internal/chezmoigit/chezmoigit.go
@@ -1,0 +1,2 @@
+// Package chezmoigit contains functions for interacting with git.
+package chezmoigit
