--- internal/git/git.go
@@ -1,2 +1,0 @@
-// Package git contains functions for interacting with git.
-package git
